Version 2019-11-07         ADC 可以使用, 但是 在VCC 上延開始測試 會不准                              RST 上延 則正常

+++ H743_Cat/Inc/var.h
@@ -49,7 +49,7 @@ extern u8 iExtCLK;
 extern u32 uInterHigh;
 
 
-extern u32    ADC_DATA[100];
+extern u16    ADC_DATA[100];
 
 extern u32    VCCEvent;
 extern u8     VCCSaved;
